1、修复如下错误。   File "/tmp/onefile_24479_1745330966_490364/app.py", line 658, in stop_endpoint AttributeError: 'NoneType' object has no attribute 'terminate'

+++ app.py
@@ -655,8 +655,9 @@ def bind_all_router():
         del_id_list=[]
         for id,v in task_status.items():
             process = v.get('process')
-            process.terminate()
-            await process.wait()
+            if process is not None:
+                process.terminate()
+                await process.wait()
             log_task_cancel(id)
             # print("after_process_terminate", othertasks)
             del_id_list.append(id)
